# openapi.yaml
openapi: 3.0.3
info:
  title: CarPool API
  version: 0.1.0
  description: API contract for the CarPool route-based ride-sharing platform.
servers:
  - url: http://localhost:8000
...
paths:
...
  /routes:
    get:
      tags: [Routes]
      summary: List routes for the logged in user
      operationId: listRoutes
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: List of routes belonging to the authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedRoutes'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Routes]
      summary: Create a route with ordered stops
      operationId: createRoute
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteCreate'
      responses:
        '201':
          description: Route created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Route' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /routes/{routeId}:
    get:
      tags: [Routes]
      summary: Get a route
      operationId: getRoute
      parameters:
        - { $ref: '#/components/parameters/RouteId' }
      responses:
        '200':
          description: Route details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Route' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    RouteId:
      name: routeId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/UUID' }
...
  responses:
    BadRequest:
      description: Bad request or validation failure.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Authentication is required or invalid.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: The user lacks the required role.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Request conflicts with the current resource state.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    UUID: { type: string, format: uuid }
    Timestamp: { type: string, format: date-time }
    Error:
      type: object
      required: [detail]
      properties:
        detail: { type: string }
        code: { type: string }
    UserRole:
      type: string
      enum: [driver, rider]
...
    LocationStatus:
      type: string
      enum: [active, inactive]
    User:
      type: object
      required: [id, name, email, roles, created_at, updated_at]
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        name: { type: string, maxLength: 100 }
        email: { type: string, format: email, maxLength: 255 }
        roles:
          type: array
          items: { $ref: '#/components/schemas/UserRole' }
        created_at: { $ref: '#/components/schemas/Timestamp' }
        updated_at: { $ref: '#/components/schemas/Timestamp' }
...
    Location:
      type: object
      required: [id, name, city, status, created_at, updated_at]
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        name: { type: string, maxLength: 255 }
        city: { type: string, maxLength: 255 }
        lat: { type: number, format: double, minimum: -90, maximum: 90, nullable: true }
        lng: { type: number, format: double, minimum: -180, maximum: 180, nullable: true }
        status: { $ref: '#/components/schemas/LocationStatus' }
        created_at: { $ref: '#/components/schemas/Timestamp' }
        updated_at: { $ref: '#/components/schemas/Timestamp' }
...
    Vehicle:
      type: object
      required: [id, driver_id, make, model, registration_number, total_seats, created_at, updated_at]
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        driver_id: { $ref: '#/components/schemas/UUID' }
        make: { type: string, minLength: 1, maxLength: 100 }
        model: { type: string, minLength: 1, maxLength: 100 }
        registration_number: { type: string, minLength: 1, maxLength: 50 }
        total_seats: { type: integer, minimum: 1, maximum: 50 }
        created_at: { $ref: '#/components/schemas/Timestamp' }
        updated_at: { $ref: '#/components/schemas/Timestamp' }
...
    RouteStop:
      type: object
      required: [id, route_id, location_id, sequence, eta_offset_minutes, created_at, updated_at]
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        route_id: { $ref: '#/components/schemas/UUID' }
        location_id: { $ref: '#/components/schemas/UUID' }
        sequence: { type: integer, minimum: 0 }
        eta_offset_minutes: { type: integer, minimum: 0 }
        location: { $ref: '#/components/schemas/Location' }
        created_at: { $ref: '#/components/schemas/Timestamp' }
        updated_at: { $ref: '#/components/schemas/Timestamp' }
    RouteStopCreate:
      type: object
      required: [location_id, sequence, eta_offset_minutes]
      properties:
        location_id: { $ref: '#/components/schemas/UUID' }
        sequence: { type: integer, minimum: 0 }
        eta_offset_minutes: { type: integer, minimum: 0 }
    Route:
      type: object
      required: [id, driver_id, vehicle_id, name, stops, created_at, updated_at]
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        driver_id: { $ref: '#/components/schemas/UUID' }
        vehicle_id: { $ref: '#/components/schemas/UUID' }
        name: { type: string, maxLength: 255 }
        driver: { $ref: '#/components/schemas/User' }
        vehicle: { $ref: '#/components/schemas/Vehicle' }
        stops:
          type: array
          minItems: 2
          items: { $ref: '#/components/schemas/RouteStop' }
        created_at: { $ref: '#/components/schemas/Timestamp' }
        updated_at: { $ref: '#/components/schemas/Timestamp' }
    RouteCreate:
      type: object
      required: [vehicle_id, name]
      properties:
        vehicle_id: { $ref: '#/components/schemas/UUID' }
        name: { type: string, minLength: 1, maxLength: 255 }
        stops:
          type: array
          minItems: 2
          items: { $ref: '#/components/schemas/RouteStopCreate' }
...
    PaginatedRoutes:
      type: object
      required: [items, page, limit, total]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Route' }
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
...